# contracts/openapi/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mutiAI Core API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/feasibility-checks/{feasibility_check_id}": {
      "get": {
        "tags": [
          "feasibility"
        ],
        "summary": "Get Feasibility Check",
        "operationId": "get_feasibility_check_api_v1_feasibility_checks__feasibility_check_id__get",
        "parameters": [
          {
            "name": "feasibility_check_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Feasibility Check Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeasibilityCheckResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{organization_id}/versions/{spec_version_id}/feasibility-checks": {
      "get": {
        "tags": [
          "feasibility"
        ],
        "summary": "List Organization Feasibility Checks",
        "operationId": "list_organization_feasibility_checks_api_v1_organizations__organization_id__versions__spec_version_id__feasibility_checks_get",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Organization Id"
            }
          },
          {
            "name": "spec_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spec Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeasibilityCheckResponse"
                  },
                  "title": "Response List Organization Feasibility Checks Api V1 Organizations  Organization Id  Versions  Spec Version Id  Feasibility Checks Get"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/feasibility-checks": {
      "get": {
        "tags": [
          "feasibility"
        ],
        "summary": "List Task Feasibility Checks",
        "operationId": "list_task_feasibility_checks_api_v1_tasks__task_id__feasibility_checks_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeasibilityCheckResponse"
                  },
                  "title": "Response List Task Feasibility Checks Api V1 Tasks  Task Id  Feasibility Checks Get"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorEnvelope"
      },
      "FeasibilityCheckResponse": {
        "properties": {
          "feasibility_check_id": {
            "type": "string",
            "title": "Feasibility Check Id"
          },
          "target_type": {
            "type": "string",
            "title": "Target Type"
          },
          "target_id": {
            "type": "string",
            "title": "Target Id"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "validator_version": {
            "type": "string",
            "title": "Validator Version"
          },
          "input_hash": {
            "type": "string",
            "title": "Input Hash"
          },
          "requirements": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Requirements"
          },
          "profile_revisions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Profile Revisions"
          },
          "outcome": {
            "$ref": "#/components/schemas/FeasibilityOutcome"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FeasibilityFindingResponse"
            },
            "type": "array",
            "title": "Findings"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "feasibility_check_id",
          "target_type",
          "target_id",
          "phase",
          "validator_version",
          "input_hash",
          "requirements",
          "profile_revisions",
          "outcome",
          "findings",
          "created_at"
        ],
        "title": "FeasibilityCheckResponse"
      },
      "FeasibilityFindingResponse": {
        "properties": {
          "reason_code": {
            "type": "string",
            "title": "Reason Code"
          },
          "role_key": {
            "type": "string",
            "title": "Role Key"
          },
          "binding_key": {
            "type": "string",
            "title": "Binding Key"
          },
          "capability": {
            "type": "string",
            "title": "Capability"
          },
          "required": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Required"
          },
          "actual": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Actual"
          },
          "alternative_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Alternative Codes"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "alternatives": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Alternatives"
          }
        },
        "type": "object",
        "required": [
          "reason_code",
          "role_key",
          "binding_key",
          "capability",
          "required",
          "actual",
          "alternative_codes",
          "message",
          "alternatives"
        ],
        "title": "FeasibilityFindingResponse"
      },
      "FeasibilityOutcome": {
        "type": "string",
        "enum": [
          "feasible",
          "conditional",
          "blocked",
          "capability_unknown"
        ],
        "title": "FeasibilityOutcome"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}